+++ src/DataLayer.EntityFramework/Repositories/BlogListRepository.cs
@@ -8,7 +8,7 @@ using AlwaysMoveForward.AnotherBlog.Common.DomainModel;
 
 namespace AlwaysMoveForward.AnotherBlog.DataLayer.Repositories
 {
-    public class BlogListRepository : EntityFrameworkRepository<BlogList, int>, IBlogListRepository
+    public class BlogListRepository : EntityFrameworkRepository<BlogList, long>, IBlogListRepository
     {
         internal BlogListRepository(IUnitOfWork unitOfWork)
             : base(unitOfWork)
@@ -20,7 +20,7 @@ namespace AlwaysMoveForward.AnotherBlog.DataLayer.Repositories
             get { return "Id"; }
         }
 
-        public IList<BlogList> GetByBlog(int blogId)
+        public IList<BlogList> GetByBlog(long blogId)
         {
             IQueryable<BlogList> dtoList = from foundItem in ((UnitOfWork)this.UnitOfWork).DataContext.BlogLists
                                            where foundItem.Blog.Id == blogId
@@ -29,14 +29,14 @@ namespace AlwaysMoveForward.AnotherBlog.DataLayer.Repositories
             return dtoList.ToList();
         }
 
-        public BlogList GetByIdAndBlogId(int listId, int blogId)
+        public BlogList GetByIdAndBlogId(long listId, long blogId)
         {
             return (from foundItem in ((UnitOfWork)this.UnitOfWork).DataContext.BlogLists
                     where foundItem.Id == listId && foundItem.Blog.Id == blogId
                     select foundItem).SingleOrDefault();
         }
 
-        public BlogList GetByNameAndBlogId(string name, int blogId)
+        public BlogList GetByNameAndBlogId(string name, long blogId)
         {
             return (from foundItem in ((UnitOfWork)this.UnitOfWork).DataContext.BlogLists
                     where foundItem.Name == name && foundItem.Blog.Id == blogId
