+++ src/Web/Areas/Admin/Models/UserModel.cs
@@ -8,7 +8,7 @@ namespace AlwaysMoveForward.AnotherBlog.Web.Areas.Admin.Models
         public AdminCommon Common { get; set; }
         public IPagedList<AnotherBlogUser> Users { get; set; }
         public IDictionary<RoleType.Id, string> Roles { get; set; }
-        public IDictionary<int, Blog> Blogs { get; set; }
+        public IDictionary<long, Blog> Blogs { get; set; }
         public AnotherBlogUser CurrentUser { get; set; }
         public IList<Blog> BlogsUserCanAccess { get; set; }
     }
